@@ -5,8 +5,7 @@ import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 
 public class NoCommand extends Command {
     @Override
-    public void getHelp(GuildMessageReceivedEvent event) {
-    }
+    public void getHelp(GuildMessageReceivedEvent event) {}
 
     @Override
     public void execute(GuildMessageReceivedEvent event) {
