@@ -12,7 +12,8 @@ class CrocodileTest {
         var crocodileTest = new Crocodile();
         crocodileTest.pathToWords = "src/test/java/com/urfu/Tamada/command/crocodile/testCrocodileWords.txt";
         crocodileTest.fillWords();
-        var a = new ArrayList<String>(){"Матанализ"};
+        var a = new ArrayList<String>();
+        a.add("Матанализ");
         assertArrayEquals(crocodileTest.words.toArray(), a.toArray());
     }
 }
