@@ -1,5 +1,6 @@
 package com.urfu.Tamada.command.database;
 
+import com.urfu.Tamada.command.database.anecdotes.Anecdotes;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -8,7 +9,7 @@ class DatabaseTest {
 
     @Test
     void getAnecdoteById() {
-        var anek = new Database().getAnecdoteById(10);
+        var anek = new Anecdotes().getAnecdoteById(10);
         assertNotEquals(anek, "");
     }
 
@@ -17,7 +18,7 @@ class DatabaseTest {
         var anek = "Учительница: \\n- Дети угадайте, как зовут всем известного персонажа, начинающегося " +
                 "\\nна букву \"Б\", живущего в стране дураков, который всюду совал свой \\nдлиный нос? " +
                 "\\n- Березовский! - дружно кричат дети.";
-        var databaseAnek = new Database().getAnecdoteById(10);
+        var databaseAnek = new Anecdotes().getAnecdoteById(10);
         assertEquals(anek, databaseAnek);
     }
 }
